Docs(README/badges): Link download stats to page bottom
The downloads graph on crates.io does not have an ID a link
can point to. However, all IDs contained in the HTML-rendered
version of the README.md are prefixed with 'user-content-'
for sanitization purposes. Use this implementation detail in
order to add a hidden element at the end of the file and point
to it in the link. Hopefully, it will work.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>