Feat(main+cargo): Filter more packages out of the version check
Closes #27: Git and path packages are made to be also considered as if
`skip-check` was enabled for them. Indeed, the version check is
currently performed on the remote registry, so any path or Git package
that is not available there will fail to version check and even if they
are there, the fetched version is not comparable to the path's or the
repo's. Instead, all these lead to the same as if their per-package
`skip-check` option was enabled as this is an easy case to handle.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>