Refactor(ship): Simplify no-fail-fast handling
Take inspiration from the implementation of the `jettison` equivalent:
there is no need to return the exit status in order to produce an error
and then wrap it conditionally. Instead, the status is checked in the
concerned functions directly and the error wrapping is done at a more
adequate place. It also fixes the error message for Binstall.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>