Refactor: Rename EffectiveConfig to EffectiveShipConfig
It is very specialized for use with the `ship` command, so adding
another command to it is not convenient or refactoring to keep a single
type for all commands seems like it would end up more bothersome than
anything else for the moment. Instead, change its name to keep things
separate as there is not much need to put things in common for now.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>